Fixed a bug in Leaderboard
